Move tests util functions to new document
